Function rust2go::new_atomic_slot

pub fn new_atomic_slot<T, A>() -> (SlotReader<T, A>, SlotWriter<T, A>)

Create a pair of SlotReader and SlotWriter. There’s 2 reasons to use it when async rust to go ffi(Go holds writer and rust holds reader):

Rust cannot guarantee trying read before go write.

Rust can dealloc the memory before go write by simply drop it if using a Box directly.